feat: Implement GEE source
